art: SketchHost + sketch-host action
Port components/art/sketch-host.tsx → Svelte 5.
sketchHost is a Svelte action that owns the full sketch lifecycle
(setup, rAF tick, ResizeObserver, teardown) on a canvas element.
SketchHost.svelte is a thin wrapper: <canvas use:sketchHost={params}>.
